+++ src/Time.h
@@ -1,0 +1,7 @@
+#pragma once
+#include <Arduino.h>
+
+void AddTime();
+void startTimer();    // 开始计时
+void pauseTimer();    // 暂停计时
+void resetTimer();    // 重置计时（清零）