+++ src/display.h
@@ -23,15 +23,27 @@ enum Page
     PAGE_FUNCTION,
     PAGE1,
     PAGE2,
-    PAGE_COUNT
+    PAGE_COUNT,
+    PAGE_DIANDENG
 };
 
-// 按钮声明（定义在 main.cpp）
-extern Button home;
+// 按钮声明（实际定义在 loop.cpp）
+// 页面
+extern Button caidan;
 extern Button page1;
 extern Button page2;
 extern Button backBtn;
+
+// 计时器
 extern Button functionBtn;
+extern Button jishikaishi;
+extern Button jishitingzhi;
+extern Button jishichongzhi;
+
+// LED 灯
+extern Button onled;
+extern Button offled;
+extern Button diandeng;
 
 // 显示
 TFT_eSPI &getTft();
@@ -56,6 +68,19 @@ void gotopage2();
 
 Page getCurrentPage();
 
+// 计时功能
+
+void jishion();
+void jishioff();
+
+// 点灯功能
+
+void drawpagediandeng();
+void switchdiandeng();
+
 // 按钮
 void drawButton(const Button &btn);
 bool isInButton(uint16_t rawX, uint16_t rawY, const Button &btn);
+
+// JPEG 图片显示
+void drawJpeg(const uint8_t *data, uint32_t size, int x, int y);