+++ src/led.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void ledon();
+void ledoff();