+++ src/loop.h
@@ -1,3 +1,9 @@
 #pragma once
 
+#include "loop.h"
+#include "display.h"
+#include "touch.h"
+#include "config.h"
+#include "Time.h"
+
 void loop();